Crate markrs

Modules

config 🔒

This module handles the configuration I/O for the application.

error 🔒

html_generator 🔒

This module provides functionality to generate HTML from markdown block elements.

io 🔒

This module provides functionality related to reading/writing files.

lexer 🔒

This module provides functionality to tokenize a line of markdown text into a vector of Token enums.

parser 🔒

This module contains the parser for converting tokenized Markdown lines into structured Markdown elements.

thread_pool 🔒

types 🔒

This module defines the types used in the markdown parser, including tokens, inline elements, block elements, and a cursor for navigating through tokens.

utils 🔒

watch 🔒

Structs

Cli 🔒

Statics

CONFIG 🔒

Functions

build 🔒

generate_static_site 🔒

main 🔒

run 🔒