markrs

Module config

This module handles the configuration I/O for the application.

Structs

Config

Represents the global configuration for the application.

HtmlConfig

Manages all configuration for HTML generation

LexerConfig

Manages all configuration for tokenization

Enums

Error

Functions

default_css 🔒

Sets the default CSS file to “default” in the case that the css_file field is omitted

default_prism_theme 🔒

Sets the default PrismJS theme to “vsc-dark-plus” in config.toml

default_tab_size 🔒

init_config

Initializes the global configuration from the specified file path

sanitize_by_default 🔒

Sets sanitize_html to true by default in config.toml

validate_config 🔒

Validates the configuration by checking if the original config file matches the filled config