markrs

Module io

This module provides functionality related to reading/writing files.

Functions

copy_css_to_output_dir

Copies a CSS file to the specified output directory.

copy_favicon_to_output_dir

Copies a favicon file to the specified output directory.

copy_file_to_output_dir

Copies a file from the input path to the specified output directory, optionally creating a subdirectory.

copy_image_to_output_dir

Copies an image file to the specified output directory.

does_config_exist

Checks if the configuration file exists at the specified path.

get_config_path

Returns the OS-specific configuration path.

read_file

Reads the contents of a file into a String.

read_input_dir

Reads all markdown files from the specified input directory and returns their contents.

visit_dir 🔒

Helper function to recursively visit subdirectories and collect markdown file contents.

write_default_config

Writes the default configuration to the configuration file to the OS-specific default configuration path.

write_default_css_file

Writes a default CSS file to the specified output directory.

write_html_to_file

Writes the provided HTML string to a file in the specified output directory.