markrs::parser

Function group_lines_with_leading_whitespace

fn group_lines_with_leading_whitespace(
    blocks: &mut Vec<Vec<Token>>,
    current_block: &mut Vec<Token>,
    previous_block: &mut Vec<Token>,
    line: &[Token],
)

Groups lines with leading whitespace into blocks based on the previous block’s content.

Arguments

blocks - A mutable reference to a vector of blocks, where each block is a vector of tokens.

current_block - A mutable reference to the current block being processed.

previous_block - A mutable reference to the previous block, used for context.

line - A mutable reference to the current line being processed, which is a vector of tokens.