markrs::parser

Function group_tabbed_lines

fn group_tabbed_lines(
    blocks: &mut Vec<Vec<Token>>,
    current_block: &mut Vec<Token>,
    previous_block: &mut Vec<Token>,
    line: &[Token],
)

Groups tabbed lines into blocks based on the previous block’s content.

Note that this function short-circuits when the first token of the line is a raw HTML tag, to allow for indented HTML.

Arguments

blocks - A mutable reference to a vector of blocks, where each block is a vector of tokens.

current_block - A mutable reference to the current block being processed.

previous_block - A mutable reference to the previous block, used for context.

line - A mutable reference to the current line being processed, which is a vector of tokens.