markrs

Module utils

Functions

build_rel_prefix

Builds a relative prefix path based on the provided HTML relative path. Note that this function does not add the final file name, only the preceding directories.

push_buffer_to_collection

Utility function for pushing a String buffer to a generic collection.